requirements.txt
flask
gevent
opencv-python
pillow
scipy
face_recognition
